Small cleanups
Adding collectors created a race condition; if two packets from
the same package were processed simultaneously, they could both
be created and overwrite eachother. Fixed by wrapping the check
if exists and create if not in one mutex lock and getting the
lock on the collector before releasing the collectors lock.

Cleaned up the run logic to compute how long it should sleep.

Made a newCollector function to pull that logic out of the Receive
method.